requirements.txt
randomcolor~=0.4.4.6
pillow~=12.0.0
requests~=2.32.5